Add support for tasks with fixed duration
Currently, we support tasks with just a description, deadline tasks with
a deadline and event with a start and end date.

There is however no support for tasks which just have a fixed duration
but no specified start or end date. For example, one might want to keep
track of a project to work on that he estimates will take 2 hours, but
have not decided when to actually start it.

To fix this, we add a new command that allows user to add a new type of
task -- fixed duration task. These are tasks with a fixed duration
specified but no start or end date specified.

It is done so, as this is in accordance with the way other tasks have
been created, reducing the need to guide users on how to use the new
command.