Use streams
Currently, the help message for date time usage, leaves a newline at the
end of its message.

This causes the gui to not render the message nicely.

Fix the newline addition at the end. At the same time, get to use java
streams (even though other part of the codebase already previously use
streams).

This makes the gui render the message more nicely.