Remove ascii art in greet message
Ascii art of yappy is displayed in the greet message

The ascii art formatting does not seem to be font agnostic. The ascii is
verified to display properly on a Linux machine with JetBrains Mono
font, but is unaligned on a Windows machine using the default system
font. This is likely due to Windows font not being monospaced. This make
the ascii art not rendered uniformly across platforms, resulting in
misalignment and hence distortion of the logo on some platforms.

To fix, remove the ascii art from greet message altogether.

Doing so, will eliminate this non-uniform experience across platforms.
The ascii can be removed as it is not important to the functionality of the
program nor does it add much to the aesthetics of it.